integration_tests/test-requirements.txt
git+https://github.com/wazo-pbx/xivo-test-helpers.git
git+https://github.com/wazo-pbx/xivo-lib-rest-client.git
git+https://github.com/wazo-pbx/wazo-provd-client.git
docker-compose
docker
nose
pyhamcrest
requests
